price endpoint setup
